Add Material icons in edit mode
1- This adds a new Material icon to identify the material entities that are linked to a parent.
Since those material entities become invisible, this help to figure their presence.

2- The Create App - Preference: "Show Zones in Create Mode"
has been removed, since it was doing nothing for a couple of years now.
the method behind is still in the API, but do nothing. 
(abandoned since it was only drawing a box that wasn't useful to figure the shape of the zone)
There was no real value to bring this back in a different way.

3- The Create App - Preference: "Show Lights and Particle Systems in Create Mode"
has been renamed for "Show Icons in Create Mode"
since this setting was also managing all the icons (including Zones and now the Parented Materials)
The objective of the setting is mainly to simplify the display at the edit when a scene is very dense.
(In the future, we might want to make this "à la carte". For now, I think we can try as is. )