Avatar auto scale: match avatar eye height to user eye height.
This adds an 'Auto' button to the avatar scale slider.  The Auto button scales the avatar so that its eye height matches the user's real-world eye height (based on the 'User real world height' setting).  In this way, the world is displayed at exactly 100% scale to the user in VR.

The '1x' button now sets the avatar scale to 1.0 even if the current scale is at the same 'notch' as 1.0 on the scale slider (previously it had no effect in that situation).

Test done:
- In VR, kneel to position your head as if you were 1m tall.
- Set 'User real-world height' to 1m (in Settings > Controls).
- Open the Avatar app and click on the sliders icon at the top-right.
- Click the '1x' button and observe that the world is miniaturised.
- Click the 'Auto' button and observe that the world is displayed at actual size.