Remove MyAvatar "squat" recenter from ad6bbc7ff6954461c35da6ddbdef7b71bf7c5c55
The "squat" recenter pre-dates the sit/stand detection, but the current version basically came from ad6bbc7ff6954461c35da6ddbdef7b71bf7c5c55 in 2018.  I'm removing it because it doesn't seem useful anymore (with or without the other changes in this branch).

What it did (on the current master) :
- If "Avatar leaning behavior" was "Auto"
- and the user was standing
- but their head was more than 5cm below standing height
- and their chest was upright
- for 30 seconds
- do a vertical recenter so that the avatar stands straight (instead of having knees bent).
When the user stood up straight after that recenter, the avatar would hover off the ground until the next recenter (eg. walking).