New Menubar
Add a New Menubar in the create app, just over the toolbar.
The Delete Red button is now back in the toolbar. 
The radius search label has been replaced by a radar icon, the space has been recuperated for the menubar.
So we have now space for more menu & menu items, and for more button too. which will help for future addition to this application.