Fix copying jsdoc to interface
Take the jsdoc from tools/jsdoc and copy it to interface.

A subtle issue was found: It turns out that for cmake it matters
what gets parsed first, especially if you rely on setting a variable
in one project and reading it in another.

This can result in things like builds that fail the first time, then
work after re-running cmake.

Generation of jsdoc and screenshare were moved above interface,
because they're a dependency of it.