Fix for Issue #1517
Fix for Issue #1517
Locked entities were simply ignored and it was still processing their children.
Now it's no more possible to duplicate if the selection and their children contains a locked entities.
It was also unable to add the new parent if the duplicated child entity was locked. Which was causing some other issues.