Replace "Explore" by "Places" application.
This adds "Places" application as a default application.
This removes the "Explore" application from default application. 

However, the "Explore" application is still available in "Edit > Running Scripts..." 
under "communityScripts" directory, for people that would like to create "beacons". 

Note that the beacon's placenames are listed in the "Places" application, 
but the feature to create new ones has been left behind because that solution had serious flaws.