Rebranded Resources for the installer
This rebrands the resource used for the installer. 

Note 1: I have not upgrade the file under "launchers" directory, these are visibly not used (still High Fidelity brand)

Note 2: There is one file I can't modify: "install-folder.rsrc". There is nothing to edit that on the web... this would need a mac.

Note 3: there was a slogan on 2 pictures under the logo: "virtual realities for one, for all". I left this blank for now... We would need a slogan to replace tit.